requirements.txt
django
djangorestframework
django-cors-headers
cryptography
python-dotenv
aiohttp
asyncio